less 6

+++ app/schemas.py
@@ -1,12 +1,14 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field, field_validator
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=127)
-    amount: float
+    amount: Decimal
     description: str | None = Field(None, max_length=255)
 
     @field_validator('amount')
-    def amount_must_be_posititve(cls, v: float) -> float:
+    def amount_must_be_posititve(cls, v: Decimal) -> Decimal:
         #Check amount is positive
         if v <= 0:
             raise ValueError("Amount must be positive")
@@ -21,7 +23,7 @@ class OperationRequest(BaseModel):
 
 class CrateWalletRequest(BaseModel):
     name: str = Field(..., max_length=127)
-    initial_balance: float = 0
+    initial_balance: Decimal = 0
 
     @field_validator("name")
     def name_not_empty(cls, v: str) -> str:
@@ -31,7 +33,7 @@ class CrateWalletRequest(BaseModel):
         return v
     
     @field_validator('initial_balance')
-    def balance_not_negative(cls, v: float) -> float:
+    def balance_not_negative(cls, v: Decimal) -> Decimal:
         #Check initial_balace is positive
         if v < 0:
             raise ValueError("Initial balace cannot be negative")
